added context at the start
love this page! added context at the start to be able to get right in. 
I remove 'simply' and 'easily' because there is always one person for whom it isn't and it's a bit discouraging. in this case I wasn't getting what was going on immediately myself...
love the meta MCE - immediatly tought, well, why 100 and not 1000? ;)